src/dashboard/requirements.txt
dash
plotly
pandas
numpy
folium
pillow
elasticsearch
boto3
python-dotenv
datetime